+++ lib/interceptor.js
@@ -417,12 +417,12 @@ Interceptor.prototype.matchAddress = function matchAddress(options) {
   return comparisonKey === matchKey
 }
 
-Interceptor.prototype.filteringPath = function filteringPath() {
-  if (_.isFunction(arguments[0])) {
-    this.scope.transformFunction = arguments[0]
-  }
+Interceptor.prototype.filteringPath = function filteringPath(...args) {
+  this.scope.filteringPath(...args)
   return this
 }
+
+// filtering by path is valid on the intercept level, but not filtering by request body?
 
 Interceptor.prototype.discard = function discard() {
   if ((this.scope.shouldPersist() || this.counter > 0) && this.filePath) {
