+++ tests/test_scope.js
@@ -5,6 +5,7 @@ const { test } = require('tap')
 const proxyquire = require('proxyquire').noPreserveCache()
 const Interceptor = require('../lib/interceptor')
 const nock = require('..')
+const got = require('./got_client')
 
 require('./cleanup_after_each')()
 
@@ -84,5 +85,84 @@ test('loadDefs throws expected when fs is not available', t => {
 
   t.throws(() => loadDefs(), { message: 'No fs' })
 
+  t.end()
+})
+
+test('filter path with function', async t => {
+  const scope = nock('http://example.test')
+    .filteringPath(() => '/?a=2&b=1')
+    .get('/?a=2&b=1')
+    .reply(200, 'Hello World!')
+
+  const { statusCode } = await got('http://example.test/', {
+    query: { a: '1', b: '2' },
+  })
+
+  t.equal(statusCode, 200)
+  scope.done()
+})
+
+test('filter path with regexp', async t => {
+  const scope = nock('http://example.test')
+    .filteringPath(/\d/g, '3')
+    .get('/?a=3&b=3')
+    .reply(200, 'Hello World!')
+
+  const { statusCode } = await got('http://example.test/', {
+    query: { a: '1', b: '2' },
+  })
+
+  t.equal(statusCode, 200)
+  scope.done()
+})
+
+test('filteringPath with invalid argument throws expected', t => {
+  t.throws(() => nock('http://example.test').filteringPath('abc123'), {
+    message:
+      'Invalid arguments: filtering path should be a function or a regular expression',
+  })
+  t.end()
+})
+
+test('filter body with function', async t => {
+  let filteringRequestBodyCounter = 0
+
+  const scope = nock('http://example.test')
+    .filteringRequestBody(body => {
+      ++filteringRequestBodyCounter
+      t.equal(body, 'mamma mia')
+      return 'mamma tua'
+    })
+    .post('/', 'mamma tua')
+    .reply(200, 'Hello World!')
+
+  const { statusCode } = await got('http://example.test/', {
+    body: 'mamma mia',
+  })
+
+  t.equal(statusCode, 200)
+  scope.done()
+  t.equal(filteringRequestBodyCounter, 1)
+})
+
+test('filter body with regexp', async t => {
+  const scope = nock('http://example.test')
+    .filteringRequestBody(/mia/, 'nostra')
+    .post('/', 'mamma nostra')
+    .reply(200, 'Hello World!')
+
+  const { statusCode } = await got('http://example.test/', {
+    body: 'mamma mia',
+  })
+
+  t.equal(statusCode, 200)
+  scope.done()
+})
+
+test('filteringRequestBody with invalid argument throws expected', t => {
+  t.throws(() => nock('http://example.test').filteringRequestBody('abc123'), {
+    message:
+      'Invalid arguments: filtering request body should be a function or a regular expression',
+  })
   t.end()
 })
